catkin plugin: extract Wstool into its own module
Currently Wstool is a class within the Catkin plugin. However, in order
to support recursively parsing rosinstall files, Wstool needs to undergo
some changes. In order to ease that work, extract Wstool into its own
module.

Progress on #1714
