+++ snapcraft/internal/errors.py
@@ -25,7 +25,7 @@ class SnapcraftError(Exception):
     """
     fmt = 'Daughter classes should redefine this'
 
-    def __init__(self, **kwargs):
+    def __init__(self, **kwargs) -> None:
         for key, value in kwargs.items():
             setattr(self, key, value)
